[C++][CORI] Update to Cori compilation+backtrace_symbols.
Adding math.h + other mandatory headers.
2 version for staticHaskKey computation:
 MACOS unchanged (TOTEST!)
 Cori --> using regex because of diffent backtrace_symbols output.
uint32_t instead of unsigned long, because on Cori g++ does not make
the difference between uint64_t and unsigned long
(maybe long long is 128?)

+++ internal/include/DynFuncCall.hpp
@@ -46,8 +46,8 @@ class DynFuncCall
         DynFuncCall(vector<void*>, string);
         // static
         DynFuncCall(vector<void*>, string, bool);
-        DynFuncCall(vector<void*>, unsigned long); 
-        DynFuncCall(vector<void*>, unsigned long, bool);
+        DynFuncCall(vector<void*>, uint32_t);
+        DynFuncCall(vector<void*>, uint32_t, bool);
         Value getJsonValue();
         vector<void*> getBtVector();
         bool applyStrategy();
